# backend/docs/openapi.yaml
components:
  schemas:
...
    ErrorResponse:
      properties:
        detail:
          title: Detail
          type: string
      required:
      - detail
      title: ErrorResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    MeetCreate:
      properties:
        data_range:
          anyOf:
          - items:
              items:
                type: string
              type: array
            type: array
          - type: 'null'
          title: Data Range
        description:
          anyOf:
          - maxLength: 400
            type: string
          - type: 'null'
          title: Description
        duration:
          anyOf:
          - type: string
          - type: 'null'
          title: Duration
        link:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Link
        name:
          maxLength: 128
          minLength: 1
          title: Name
          type: string
      required:
      - name
      - data_range
      title: MeetCreate
      type: object
    MeetResponse:
      properties:
        data_range:
          anyOf:
          - items:
              items:
                type: string
              type: array
            type: array
          - type: 'null'
          title: Data Range
        description:
          anyOf:
          - maxLength: 400
            type: string
          - type: 'null'
          title: Description
        duration:
          anyOf:
          - type: string
          - type: 'null'
          title: Duration
        hash:
          format: uuid
          title: Hash
          type: string
        isCreator:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Iscreator
        isCreatorAuth:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Iscreatorauth
        link:
          anyOf:
          - maxLength: 128
            type: string
          - type: 'null'
          title: Link
        name:
          maxLength: 128
          minLength: 1
          title: Name
          type: string
        slots:
          default: []
          items:
            $ref: '#/components/schemas/SlotsUser'
          title: Slots
          type: array
      required:
      - name
      - data_range
      - hash
      title: MeetResponse
      type: object
...
    SlotsUser:
      properties:
        name:
          maxLength: 128
          minLength: 1
          title: Name
          type: string
        slots:
          items:
            items:
              type: string
            type: array
          title: Slots
          type: array
        userId:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Userid
      required:
      - name
      - slots
      title: SlotsUser
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    HTTPBearer:
      scheme: bearer
      type: http
info:
  description: "API \u0434\u043B\u044F Termeet"
  title: Termeet API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /meet/create:
    post:
      description: "\u0421\u043E\u0437\u0434\u0430\u0435\u0442 \u0441\u043B\u043E\u0442\
        \u044B \u0434\u043B\u044F \u0432\u0441\u0442\u0440\u0435\u0447\u0438 \u043F\
        \u043E \u043E\u043F\u0440\u0435\u0434\u0435\u043B\u0435\u043D\u043D\u044B\u043C\
        \ \u0434\u043D\u044F\u043C                 \u0438 \u043F\u0440\u043E\u043C\
        \u0435\u0436\u0443\u0442\u043A\u0443 \u0432\u0440\u0435\u043C\u0435\u043D\u0438"
      operationId: create_meeting_meet_create_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F access_\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\
            \u0435\u043A /                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\
            \u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\
            \u043D\u0430"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: "\u0421\u043E\u0437\u0434\u0430\u0442\u044C \u0432\u0441\u0442\u0440\
        \u0435\u0447\u0443"
      tags:
      - Meet
  /meet/{hash}:
    get:
      description: "\u0412\u043E\u0437\u0432\u0440\u0430\u0449\u0430\u0435\u0442 \u0441\
        \u043B\u043E\u0442\u044B \u0432\u0441\u0442\u0440\u0435\u0447\u0438 \u0434\
        \u043B\u044F \u0432\u044B\u0431\u043E\u0440\u0430,                 \u0430\
        \ \u0442\u0430\u043A\u0436\u0435 \u0443\u0436\u0435 \u0432\u044B\u0431\u0440\
        \u0430\u043D\u043D\u044B\u0435 \u0443\u0447\u0430\u0441\u0442\u043D\u0438\u043A\
        \u0430\u043C\u0438"
      operationId: get_meeting_meet__hash__get
      parameters:
      - in: path
        name: hash
        required: true
        schema:
          format: uuid
          title: Hash
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetResponse'
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F access_\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\
            \u0435\u043A /                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\
            \u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\
            \u043D\u0430"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: "\u041F\u043E\u043B\u0443\u0447\u0438\u0442\u044C \u0432\u0441\u044E\
        \ \u0438\u043D\u0444\u043E\u0440\u043C\u0430\u0446\u0438\u044E \u043E \u0432\
        \u0441\u0442\u0440\u0435\u0447\u0435"
      tags:
      - Meet
    patch:
      description: "\u041E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442 \u043F\
        \u043E\u043B\u043D\u043E\u0435 \u043E\u043F\u0438\u0441\u0430\u043D\u0438\u0435\
        \ \u0432\u0441\u0442\u0440\u0435\u0447\u0438,         \u0442\u043E \u0435\u0441\
        \u0442\u044C \u0438\u0437\u043C\u0435\u043D\u0435\u043D\u043D\u044B\u0435\
        \ \u0438 \u043D\u0435 \u0438\u0437\u043C\u0435\u043D\u0435\u043D\u043D\u044B\
        \u0435"
      operationId: edit_meeting_meet__hash__patch
      parameters:
      - in: path
        name: hash
        required: true
        schema:
          format: uuid
          title: Hash
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MeetCreate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F access_\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\
            \u0435\u043A /                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\
            \u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\
            \u043D\u0430"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D /                 \u041F\u043E\u043B\u044C\u0437\u043E\u0432\
            \u0430\u0442\u0435\u043B\u044C \u043D\u0435 \u044F\u0432\u043B\u044F\u0435\
            \u0442\u0441\u044F \u0441\u043E\u0437\u0434\u0430\u0442\u0435\u043B\u0435\
            \u043C \u0432\u0441\u0442\u0440\u0435\u0447\u0438"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: "\u0420\u0435\u0434\u0430\u043A\u0442\u0438\u0440\u043E\u0432\u0430\
        \u0442\u044C \u0432\u0441\u0442\u0440\u0435\u0447\u0443"
      tags:
      - Meet
  /meet/{hash}/slots:
    patch:
      description: "\u041E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442 \u0441\
        \u043B\u043E\u0442\u044B \u0432\u0441\u0442\u0440\u0435\u0447\u0438, \u043A\
        \u043E\u0442\u043E\u0440\u044B\u0435 \u0432\u044B\u0431\u0440\u0430\u043B\
        \                 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
        \u043B\u044C"
      operationId: add_slots_meet__hash__slots_patch
      parameters:
      - in: path
        name: hash
        required: true
        schema:
          format: uuid
          title: Hash
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlotsUser'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F access_\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\
            \u0435\u043A /                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\
            \u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\
            \u043D\u0430"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D /                 \u041F\u043E\u043B\u044C\u0437\u043E\u0432\
            \u0430\u0442\u0435\u043B\u044C \u0443\u0436\u0435 \u0434\u043E\u0431\u0430\
            \u0432\u0438\u043B \u0441\u043B\u043E\u0442\u044B \u0434\u043B\u044F \u044D\
            \u0442\u043E\u0439 \u0432\u0441\u0442\u0440\u0435\u0447\u0438 /      \
            \           \u0421\u043B\u043E\u0442\u044B \u0441 \u0442\u0430\u043A\u0438\
            \u043C \u043D\u0438\u043A\u043D\u0435\u0439\u043C\u043E\u043C \u043F\u043E\
            \u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F \u0443\u0436\
            \u0435 \u0441\u0443\u0449\u0435\u0441\u0442\u0432\u0443\u044E\u0442"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: "\u041E\u0442\u043F\u0440\u0430\u0432\u0438\u0442\u044C \u0441\u043B\
        \u043E\u0442\u044B, \u0432\u044B\u0431\u0440\u0430\u043D\u043D\u044B\u0435\
        \ \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u0435\
        \u043C"
      tags:
      - Meet
  /meet/{hash}/slots/edit:
    patch:
      description: "\u041E\u0442\u043F\u0440\u0430\u0432\u043B\u044F\u0435\u0442 \u043E\
        \u0442\u0440\u0435\u0434\u0430\u043A\u0442\u0438\u0440\u043E\u0432\u0430\u043D\
        \u043D\u044B\u0435 \u0441\u043B\u043E\u0442\u044B \u0432\u0441\u0442\u0440\
        \u0435\u0447\u0438,                 \u043A\u043E\u0442\u043E\u0440\u044B\u0435\
        \ \u0432\u044B\u0431\u0440\u0430\u043B \u043F\u043E\u043B\u044C\u0437\u043E\
        \u0432\u0430\u0442\u0435\u043B\u044C"
      operationId: edit_slots_meet__hash__slots_edit_patch
      parameters:
      - in: path
        name: hash
        required: true
        schema:
          format: uuid
          title: Hash
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlotsUser'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F access_\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\
            \u0435\u043A /                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\
            \u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\
            \u043D\u0430 /                 \u041F\u043E\u043B\u044C\u0437\u043E\u0432\
            \u0430\u0442\u0435\u043B\u044C \u0434\u043E\u043B\u0436\u0435\u043D \u0431\
            \u044B\u0442\u044C \u0430\u0443\u0442\u0435\u043D\u0442\u0438\u0444\u0438\
            \u0446\u0438\u0440\u043E\u0432\u0430\u043D,                 \u0447\u0442\
            \u043E\u0431\u044B \u0440\u0435\u0434\u0430\u043A\u0442\u0438\u0440\u043E\
            \u0432\u0430\u0442\u044C \u0441\u043B\u043E\u0442\u044B"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D /                 \u041F\u043E\u043B\u044C\u0437\u043E\u0432\
            \u0430\u0442\u0435\u043B\u044C \u043D\u0435 \u044F\u0432\u043B\u044F\u0435\
            \u0442\u0441\u044F \u0443\u0447\u0430\u0441\u0442\u043D\u0438\u043A\u043E\
            \u043C \u0432\u0441\u0442\u0440\u0435\u0447\u0438 /                 \u0421\
            \u043B\u043E\u0442\u044B \u0434\u043B\u044F \u044D\u0442\u043E\u0433\u043E\
            \ \u043F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044F\
            \ \u043D\u0435 \u043D\u0430\u0439\u0434\u0435\u043D\u044B"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: "\u041E\u0442\u0440\u0435\u0434\u0430\u043A\u0442\u0438\u0440\u043E\
        \u0432\u0430\u0442\u044C \u0441\u043B\u043E\u0442\u044B, \u0432\u044B\u0431\
        \u0440\u0430\u043D\u043D\u044B\u0435 \u043F\u043E\u043B\u044C\u0437\u043E\u0432\
        \u0430\u0442\u0435\u043B\u0435\u043C"
      tags:
      - Meet
  /meet/{hash}/slots/{username}:
    delete:
      description: "\u0423\u0434\u0430\u043B\u044F\u0435\u0442 \u0441\u043B\u043E\u0442\
        \u044B \u0432\u0441\u0442\u0440\u0435\u0447\u0438, \u043A\u043E\u0442\u043E\
        \u0440\u044B\u0435 \u0432\u044B\u0431\u0440\u0430\u043B                 \u043F\
        \u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u044C"
      operationId: delete_slots_meet__hash__slots__username__delete
      parameters:
      - in: path
        name: hash
        required: true
        schema:
          format: uuid
          title: Hash
          type: string
      - in: path
        name: username
        required: true
        schema:
          title: Username
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u0421\u0440\u043E\u043A \u0434\u0435\u0439\u0441\u0442\u0432\
            \u0438\u044F access_\u0442\u043E\u043A\u0435\u043D\u0430 \u0438\u0441\u0442\
            \u0435\u043A /                 \u041D\u0435 \u043F\u0440\u0430\u0432\u0438\
            \u043B\u044C\u043D\u044B\u0439 \u0442\u0438\u043F \u0442\u043E\u043A\u0435\
            \u043D\u0430"
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041F\u043E\u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\
            \u043B\u044C \u0437\u0430\u0431\u043B\u043E\u043A\u0438\u0440\u043E\u0432\
            \u0430\u043D /                 \u041F\u043E\u043B\u044C\u0437\u043E\u0432\
            \u0430\u0442\u0435\u043B\u044C \u043D\u0435 \u044F\u0432\u043B\u044F\u0435\
            \u0442\u0441\u044F \u0441\u043E\u0437\u0434\u0430\u0442\u0435\u043B\u0435\
            \u043C \u0432\u0441\u0442\u0440\u0435\u0447\u0438"
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: "\u041E\u0431\u044A\u0435\u043A\u0442 \u043D\u0435 \u043D\u0430\
            \u0439\u0434\u0435\u043D /             \u0421\u043B\u043E\u0442\u044B\
            \ \u0434\u043B\u044F \u044D\u0442\u043E\u0433\u043E \u043D\u0438\u043A\
            \u043D\u0435\u0439\u043C\u0430 \u043D\u0435 \u043D\u0430\u0439\u0434\u0435\
            \u043D\u044B"
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - HTTPBearer: []
      summary: "\u0423\u0434\u0430\u043B\u0438\u0442\u044C \u0441\u043B\u043E\u0442\
        \u044B, \u0432\u044B\u0431\u0440\u0430\u043D\u043D\u044B\u0435 \u043F\u043E\
        \u043B\u044C\u0437\u043E\u0432\u0430\u0442\u0435\u043B\u0435\u043C"
      tags:
      - Meet
...